stream: always emit drain if write returns false
If a write is above the high water mark but _write still manages to
fully send it synchronously, _writableState.length can be adjusted down
to 0 synchronously with the write that returns false, but 'drain' will
not be emitted until process.nextTick. Ensure that in this case, an
interleaved short write won't clear the needDrain flag.

This is especially relevant if you pipe a legacy Stream to this
writable; the legacy Stream will be paused as soon as one call to
write() returns false, and without this fix it's possible that the
legacy Stream will never be resumed.